Stepwise forward regression

Build regression model from a set of candidate predictor variables by entering predictors based on p values, in a stepwise manner until there is no variable left to enter any more.

blr_step_p_forward(model, ...)

# S3 method for default
blr_step_p_forward(model, penter = 0.3, details = FALSE,
  ...)

# S3 method for blr_step_p_forward
plot(x, model = NA, ...)

Arguments

model
An object of class lm; the model should include all candidate predictor variables.

...
Other arguments.

penter
p value; variables with p value less than penter will enter into the model

details
Logical; if TRUE, will print the regression result at each step.

x
An object of class blr_step_p_forward.

Value

blr_step_p_forward returns an object of class "blr_step_p_forward". An object of class "blr_step_p_forward" is a list containing the following components:

steps

number of steps

predictors

variables added to the model

rsquare

coefficient of determination

aic

akaike information criteria

sbc

bayesian information criteria

sbic

sawa's bayesian information criteria

adjr

adjusted r-square

rmse

root mean square error

mallows_cp

mallow's Cp

indvar

predictors

Deprecated Function

blr_step_forward() has been deprecated. Instead use blr_step_p_forward().

References

Chatterjee, Samprit and Hadi, Ali. Regression Analysis by Example. 5th ed. N.p.: John Wiley & Sons, 2012. Print.

Kutner, MH, Nachtscheim CJ, Neter J and Li W., 2004, Applied Linear Statistical Models (5th edition). Chicago, IL., McGraw Hill/Irwin.

See also

Other variable selection procedures: blr_step_aic_backward, blr_step_aic_both, blr_step_aic_forward

Examples

# stepwise forward regression
model <- glm(honcomp ~ female + read + science, data = hsb2,
  family = binomial(link = 'logit'))
blr_step_p_forward(model)

#> Forward Selection Method    
#> ---------------------------
#> 
#> Candidate Terms: 
#> 
#> 1. female 
#> 2. read 
#> 3. science 
#> 
#> We are selecting variables based on p value...
#> 
#> Variables Entered: 
#> 
#> <U+2714> read 
#> <U+2714> female 
#> <U+2714> science 
#> 
#> 
#> Final Model Output 
#> ------------------
#> 
#> <U+2714> Creating model overview. 
#> <U+2714> Creating response profile. 
#> <U+2714> Extracting maximum likelihood estimates. 
#> <U+2714> Estimating concordant and discordant pairs. 
#> 
#>                              Model Overview                              
#> ------------------------------------------------------------------------
#> Data Set    Resp Var    Obs.    Df. Model    Df. Residual    Convergence 
#> ------------------------------------------------------------------------
#>   data      honcomp     200        199           196            TRUE     
#> ------------------------------------------------------------------------
#> 
#>                     Response Summary                     
#> --------------------------------------------------------
#> Outcome        Frequency        Outcome        Frequency 
#> --------------------------------------------------------
#>    0              147              1              53     
#> --------------------------------------------------------
#> 
#>                   Maximum Likelihood Estimates                    
#> -----------------------------------------------------------------
#>  Parameter     DF    Estimate    Std. Error    z value    Pr(>|z|) 
#> -----------------------------------------------------------------
#> (Intercept)    1     -12.7772       1.9755    -6.4677      0.0000 
#>    read        1      0.1035        0.0258     4.0186       1e-04 
#>   female1      1      1.4825        0.4474     3.3139       9e-04 
#>   science      1      0.0948        0.0305     3.1129      0.0019 
#> -----------------------------------------------------------------
#> 
#>  Association of Predicted Probabilities and Observed Responses  
#> ---------------------------------------------------------------
#> % Concordant          0.8561          Somers' D        0.7147   
#> % Discordant          0.1425          Gamma            0.7136   
#> % Tied                0.0014          Tau-a            0.2794   
#> Pairs                  7791           c                0.8568   
#> ---------------------------------------------------------------
#> 

#> 
#>                  Selection Summary                   
#> ----------------------------------------------------
#>         Variable                                        
#> Step    Entered       AIC         BIC       Deviance    
#> ----------------------------------------------------
#>    1    read        183.0630    189.6596    179.0630    
#>    2    female      176.8874    186.7824    170.8874    
#>    3    science     168.2364    181.4296    160.2364    
#> ----------------------------------------------------

# stepwise forward regression plot
model <- glm(honcomp ~ female + read + science, data = hsb2,
  family = binomial(link = 'logit'))
k <- blr_step_p_forward(model)

#> Forward Selection Method    
#> ---------------------------
#> 
#> Candidate Terms: 
#> 
#> 1. female 
#> 2. read 
#> 3. science 
#> 
#> We are selecting variables based on p value...
#> 
#> Variables Entered: 
#> 
#> <U+2714> read 
#> <U+2714> female 
#> <U+2714> science 
#> 
#> 
#> Final Model Output 
#> ------------------
#> 
#> <U+2714> Creating model overview. 
#> <U+2714> Creating response profile. 
#> <U+2714> Extracting maximum likelihood estimates. 
#> <U+2714> Estimating concordant and discordant pairs. 
#> 
#>                              Model Overview                              
#> ------------------------------------------------------------------------
#> Data Set    Resp Var    Obs.    Df. Model    Df. Residual    Convergence 
#> ------------------------------------------------------------------------
#>   data      honcomp     200        199           196            TRUE     
#> ------------------------------------------------------------------------
#> 
#>                     Response Summary                     
#> --------------------------------------------------------
#> Outcome        Frequency        Outcome        Frequency 
#> --------------------------------------------------------
#>    0              147              1              53     
#> --------------------------------------------------------
#> 
#>                   Maximum Likelihood Estimates                    
#> -----------------------------------------------------------------
#>  Parameter     DF    Estimate    Std. Error    z value    Pr(>|z|) 
#> -----------------------------------------------------------------
#> (Intercept)    1     -12.7772       1.9755    -6.4677      0.0000 
#>    read        1      0.1035        0.0258     4.0186       1e-04 
#>   female1      1      1.4825        0.4474     3.3139       9e-04 
#>   science      1      0.0948        0.0305     3.1129      0.0019 
#> -----------------------------------------------------------------
#> 
#>  Association of Predicted Probabilities and Observed Responses  
#> ---------------------------------------------------------------
#> % Concordant          0.8561          Somers' D        0.7147   
#> % Discordant          0.1425          Gamma            0.7136   
#> % Tied                0.0014          Tau-a            0.2794   
#> Pairs                  7791           c                0.8568   
#> ---------------------------------------------------------------
#> 

plot(k)

#> Error in plot_stepwise(d4, "AIC"): could not find function "plot_stepwise"