Bank marketing data set

The data is related with direct marketing campaigns of a Portuguese banking institution. The marketing campaigns were based on phone calls. Often, more than one contact to the same client was required, in order to access if the product (bank term deposit) would be ('yes') or not ('no') subscribed.

Usage

bank_marketing

Format

A tibble with 4521 rows and 17 variables:

age

age of the client

job

type of job

marital

marital status

education

education level of the client

default

has credit in default?

housing

has housing loan?

loan

has personal loan?

contact

contact communication type

month

last contact month of year

day_of_week

last contact day of the week

duration

last contact duration, in seconds

campaign

number of contacts performed during this campaign and for this client

pdays

number of days that passed by after the client was last contacted from a previous campaign

previous

number of contacts performed before this campaign and for this clien

poutcome

outcome of the previous marketing campaign

y

has the client subscribed a term deposit?

Source

[Moro et al., 2014] S. Moro, P. Cortez and P. Rita. A Data-Driven Approach to Predict the Success of Bank Telemarketing. Decision Support Systems, Elsevier, 62:22-31, June 2014